+++ db/migrate/20130801183544_add_file_uploads_to_preceptors.rb
@@ -1,0 +1,6 @@
+class AddFileUploadsToPreceptors < ActiveRecord::Migration
+  def change
+    add_column :preceptors, :biosketch, :string
+    add_column :preceptors, :curriculum_vitae, :string
+  end
+end